more win32 compilation fixes

+++ sql/backends/monet5/vaults/parquet/pqc_filemetadata.c
@@ -248,7 +248,7 @@ pqc_open( pqc_file **PQ, char *fn)
 		pqc_destroy(pq);
 		return -1;
 	}
-	pq->sz = stb.st_size;
+	pq->sz = (size_t)stb.st_size;
 	pq->buffer = NULL;
 	pq->bsz = 0;
